Interface movida para o index
-- A interface Person foi motiva para o index da person
-- Corrigido imports da interface para usar 'import type'
