+++ src/userApi/user.ts
@@ -1,8 +1,8 @@
 import { sha512 } from 'js-sha512'
 import { ObjectId, Decimal128 } from 'mongodb'
 import * as currencyApi from '../currencyApi'
-import { Person } from '../db/models/person/interface'
 import { Pending } from '../db/models/person/currencies/interface'
+import type { Person } from '../db/models/person'
 import type { SuportedCurrencies as SC } from '../currencyApi'
 
 /**
